Key-valueAPI/application/api/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
redis==5.0.1
psycopg2-binary==2.9.9
pika==1.3.2
pydantic==2.5.2
pytz==2023.3
pytest==7.4.4
httpx==0.25.2
pytest-mock==3.12.0 